+++ src/main/java/com/example/rentalSystem/domain/rentalhistory/controller/RentalController.java
@@ -1,8 +1,7 @@
 package com.example.rentalSystem.domain.rentalhistory.controller;
 
 import com.example.rentalSystem.domain.member.entity.CustomerDetails;
-import com.example.rentalSystem.domain.rentalhistory.dto.CreateRentalRequest;
-import com.example.rentalSystem.domain.rentalhistory.dto.response.RentalHistoryDetailResponseDto;
+import com.example.rentalSystem.domain.rentalhistory.dto.request.CreateRentalRequest;
 import com.example.rentalSystem.domain.rentalhistory.dto.response.RentalHistoryResponseDto;
 import com.example.rentalSystem.domain.rentalhistory.service.RentalService;
 import com.example.rentalSystem.global.response.ApiResponse;
@@ -15,7 +14,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
