+++ src/main/java/com/example/rentalSystem/domain/rentalhistory/implement/RentalScheduler.java
@@ -1,21 +1,18 @@
 package com.example.rentalSystem.domain.rentalhistory.implement;
 
-import static com.example.rentalSystem.domain.facility.entity.RentalStatus.AVAILABLE;
+import static com.example.rentalSystem.domain.facility.entity.timeTable.TimeStatus.AVAILABLE;
 
 import com.example.rentalSystem.domain.facility.entity.Facility;
-import com.example.rentalSystem.domain.facility.entity.RentalStatus;
-import com.example.rentalSystem.domain.facility.entity.TimeTable;
+import com.example.rentalSystem.domain.facility.entity.timeTable.TimeStatus;
+import com.example.rentalSystem.domain.facility.entity.timeTable.TimeTable;
 import com.example.rentalSystem.domain.facility.reposiotry.TimeTableRepository;
 import com.example.rentalSystem.global.exception.custom.CustomException;
 import com.example.rentalSystem.global.response.ErrorType;
-import com.example.rentalSystem.global.response.SuccessType;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.LinkedHashMap;
-import java.util.Map.Entry;
 import lombok.RequiredArgsConstructor;
-import org.springframework.cglib.core.Local;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -37,7 +34,7 @@ public class RentalScheduler {
                 startDate)
             .orElseThrow(() -> new CustomException(ErrorType.ENTITY_NOT_FOUND));
 
-        LinkedHashMap<LocalTime, RentalStatus> timeSlot = timeTable.getTimeSlot();
+        LinkedHashMap<LocalTime, TimeStatus> timeSlot = timeTable.getTimeSlot();
 
         while (startTIme.isBefore(endTime)) {
             if (!timeSlot.get(startTIme).equals(AVAILABLE)) {
